feat: shop interface components implemented
The header received a search bar, which user can filter by choice category.
When the page load, all products appear in the showcase.
The product card received a button to user select and drop it in the cart.
The cart display the total sale and have a button to clear all items.